pyproject.toml
[tool.poetry]
name = "propleads-pro"
version = "0.1.0"
description = "Motivated seller lead generation — CrewAI Flow"
authors = ["PropLeads Pro"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = {version = ">=0.152.0,<1.0.0", extras = ["tools"]}
supabase = ">=2.4.1"
sendgrid = ">=6.11.0"
jinja2 = ">=3.1.4"
requests = ">=2.31.0"
python-dotenv = ">=1.0.1"
flask = ">=3.0.3"
pydantic = ">=2.0.0"

[tool.poetry.scripts]
propleads     = "propleads_pro.main:run"
propleads_all = "propleads_pro.main:run_all"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.crewai]
type = "flow"
